+++ src/App.js
@@ -1,5 +1,5 @@
 import React from "react";
-import "./App.css";
+import "App.css";
 
 class App extends React.Component {
   constructor(props) {
