+++ src/components/Pages/home/home.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+function Home() {
+  return <div>Home</div>;
+}
+
+export default Home;
