+++ src/components/Pages/news/news.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+function News() {
+  return <div>News</div>;
+}
+
+export default News;
