+++ src/components/Pages/policy/policy.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+function Policy() {
+  return <div>Policy</div>;
+}
+
+export default Policy;
