+++ src/components/header/mainTitle/mainTitle.css
