+++ src/components/header/navbar/sideBar/sideBar.js
@@ -2,9 +2,9 @@ import React, { useState } from "react";
 import * as FaIcons from "react-icons/fa";
 import * as AiIcons from "react-icons/ai";
 import { Link } from "react-router-dom";
-import "./sideBar.css";
+import "components/header/navbar/sideBar/sideBar.css";
 import { IconContext } from "react-icons";
-import { navbarData } from "../data/navbarData.js";
+import { navbarData } from "components/header/navbar/data/navbarData.js";
 
 const SideBar = () => {
   const [sidebar, setSidebar] = useState(false);
