--- src/components/homepage/header/navbar/reports.js
@@ -1,10 +1,0 @@
-import React from "react";
-
-function Reports() {
-  return (
-    <div className="reports">
-      <h1>Reports</h1>
-    </div>
-  );
-}
-export default Reports;
