+++ src/index.js
@@ -1,8 +1,8 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import "./index.css";
+import "index.css";
 // import App from "./App";
-import MainContainer from "./components/mainContainer.js";
+import MainContainer from "components/mainContainer.js";
 
 ReactDOM.render(
   <React.StrictMode>
